wip(slice02): S5 split roots into DesktopHostProviders/ProductProviderRoot/ProductLifecycleRoot; delete AppProviders
